Change the DjangoListField default fields to None (from "__all__")
By defaulting this to "__all__" this prevented exclude from working,
as an error is raised in the base class if both fields and exclude are
specified.

Changed the example Owner model to add a pets field.
Changed OwnerType to use exclude rather than include so the exclude option would be tested implicitly.